Start implementing a simple template for the new picker
